+++ packages/core/tools/utils/constants.js
@@ -3,7 +3,7 @@ const { join } = require('path');
 const BLOCKLIST = ['@realsystem/core'];
 
 const PACKAGES_ROOT_PATH = join(__dirname, '../../../..');
-const CORE_BUNDLE_PATH = join(__dirname, '../');
+const CORE_BUNDLE_PATH = join(__dirname, '../../');
 const CORE_BUNDLE_INDEX_PATH = join(CORE_BUNDLE_PATH, 'src/index.ts');
 const CORE_BUNDLE_PACKAGE_PATH = join(CORE_BUNDLE_PATH, 'package.json');
 const CORE_BUNDLE_OUTPUT_PATH = join(CORE_BUNDLE_PATH, 'lib/');