+++ packages/box/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @realsystem/box
 
+## 6.0.2
+
+### Patch Changes
+
+- [`e9ef3ae`](https://github.com/bigwoof91/realsystem/commit/e9ef3ae181b51d3768d05ce2a57a10e2e6ac5145) - trying to fix 'cannot find module react' error in prod
+
+- Updated dependencies [[`e9ef3ae`](https://github.com/bigwoof91/realsystem/commit/e9ef3ae181b51d3768d05ce2a57a10e2e6ac5145)]:
+  - @realsystem/styling@1.4.2
+
 ## 6.0.1
 
 ### Patch Changes